User (#18)
* initial commit for user

* change in list friends return type

* commit error in wire.go

* added main logic and di

* added test file and mockgen files

* add handler test

* [MERGE COMPLETE]fix wire.go and device_repository

* update README

* Refactor : UserID from string to uint

* Update cmd/user-svc/main.go

* Refactoring pt.2

* Refactor: Test Files

---------

+++ internal/dbmysql/notification_models.go
@@ -9,8 +9,9 @@ import (
 )
 
 type Notification struct {
-	ID            uint    `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID        uint    `gorm:"not null;index" json:"user_id"`
+	ID     uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	UserID uint64 `gorm:"not null;index" json:"user_id"`
+
 	Header        string  `gorm:"not null;size:255"`
 	Content       string  `gorm:"not null;type:text"`
 	ImageURL      *string `gorm:"size:512"`
